backend/requirements.txt
# Core Backend
flask==3.0.0
pymongo==4.5.0  # Keep your existing version

# Google Cloud Integration
google-cloud-aiplatform==1.95.1  # For Vertex AI/Gemini
google-cloud-secret-manager==2.16.1  # For secure MongoDB URI

# Data Processing (Optional)
pandas==2.0.3  # Only if needed for data cleaning
unidecode==1.3.6  # Only if processing non-ASCII text

# Environment Management
python-dotenv==1.0.0